+++ app/javascript/engine/Map.ts
@@ -956,7 +956,7 @@ export default class Map {
       if (a.unit.playerNation === b.unit.playerNation) { return 0 }
       return a.unit.playerNation === this.game?.currentPlayerNation ? -1 : 1
     })
-    if (this.game && this.game?.fireDisplaceNeeded.length > 0) {
+    if (this.game && this.game.fireDisplaceNeeded.length > 0) {
       const unit = this.game.fireDisplaceNeeded[0].unit
       const player = playerForNation(unit, this.game)
       this.game.setCurrentPlayer(player)